Add 3 unrelated figs: sorting, knots, kites/darts
The sorting diagrams are based on those in Sedgewick's algorithms book.
So far only selection sort is present.

The knot figures include
 - trefoil (2 versions)
 - whitehead links
Additionally there are 2 ways to render knots
 1. with under-crossings erased
 2. using white stripes in the center of fat lines
The macros to support these seem to be solid & versatile enough to
extend to more knots.

The drawings for Penrose's kites & darts are very preliminary. None of
it is in the portfolio yet.

Additionally, there is a document on the logic behind the diagrams:
        metapost-it-notes.tex
The kites/darts stuff is there, and eventually the write-up on yin/yang
should be moved there out of the portfolio.